move to spring-rabbitmq
