+++ rule/rabbitmq-client/src/main/java/io/opentracing/contrib/specialagent/rule/rabbitmq/client/RabbitMQAgentIntercept.java
@@ -65,32 +65,4 @@ public class RabbitMQAgentIntercept {
   public static Object enterConsume(final Object callback, final Object queue) {
     return WrapperProxy.wrap(callback, new TracingConsumer((Consumer)callback, (String)queue, GlobalTracer.get()));
   }
-
-  public static void handleDeliveryStart(Object thiz, Object props) {
-    if (WrapperProxy.isWrapper(thiz, TracingConsumer.class))
-      return;
-
-    if (AgentRuleUtil.callerEquals(1, 3, "io.opentracing.contrib.rabbitmq.TracingConsumer.handleDelivery"))
-      return;
-
-    final AMQP.BasicProperties properties = (AMQP.BasicProperties) props;
-    final Tracer tracer = GlobalTracer.get();
-
-    final Span span = TracingUtils.buildChildSpan(properties, null, tracer);
-
-    final Scope scope = tracer.activateSpan(span);
-    LocalSpanContext.set(span, scope);
-  }
-
-  public static void handleDeliveryEnd(final Throwable thrown) {
-    final LocalSpanContext context = LocalSpanContext.get();
-    if (context == null)
-      return;
-
-    if (thrown != null)
-      AgentRuleUtil.setErrorTag(context.getSpan(), thrown);
-
-    context.closeAndFinish();
-  }
-
 }
